Support simplifying the snake game example
Support making the snake game example simpler by moving generic helper functions out of the users' sight. Move generic functions to a package, based on the findings in the snake game example as documented in the explore directory.
The `ListDict` is not yet published in its own package, continue to use a `List` here for now.