Support Elm publish
Work around the 'not tagged on GitHub' error in Elm publish as reported at https://discourse.elm-lang.org/t/how-to-fix-elm-publish-not-tagged-on-github-error/3625: For using Elm publish, copy each Elm package to a dedicated repository. To support this, move the .gitignore file to the directory which defines the content of the git repository derived for Elm publish.