Simplify API based on recent learnings about Elm program types
Also, simplify the language in guides.

The new game configuration type reflects the learnings about program types from June 2021: We did not yet find a good reason to keep a dedicated type for events mid-term. Short term, the limitations in tools were reasons to have an event/'msg' type. For example, tooling in 2021 is not good at displaying or serializing functions.
For discussion of the event type, see https://discourse.elm-lang.org/t/messages-purpose/6778 and https://discourse.elm-lang.org/t/why-not-use-update-fn-state-fn-state/7037
There was also a design with the only customizable type being the event type, with the state defined as `List event`. But we did not yet find a way to make that work together with migrations, so we kept the state type instead of the event type.